Reinstate anchor pattern
This was present in @spiette's original repo. The pattern that @garethr
replaced it with isn't something that I've seen before.

It doesn't satisfy other resources chaining *before* this containing class.
They float off to the side. As demonstrated by:

https://gist.github.com/dcarley/5883522

Notifying the containing class now also notifies the service. Which required
stdlib >= 4.0, per the previous commit.